Update Linting Script and Enhance Constants for Improved Error Handling
- Changed the severity level in `lint_code.sh` from "hint" to "error" to enforce stricter linting rules, improving code quality and compliance with project standards.
- Added new constants in `onvif_constants.h` for error message buffer sizes and decimal base, enhancing clarity and maintainability in error handling.
- Updated various modules to utilize defined buffer sizes for HTTP methods, paths, and versions, ensuring consistency across the codebase.

These changes strengthen the linting process and improve error handling mechanisms, aligning with the project's goals for comprehensive testing and adherence to ONVIF standards.

+++ cross-compile/onvif/src/networking/common/epoll_server.c
@@ -61,7 +61,7 @@ void epoll_server_set_config(const struct application_config* config) {
 
   // Validate and set cleanup interval
   if (config->server.cleanup_interval >= 1 && config->server.cleanup_interval <= CLEANUP_INTERVAL_MAX_SEC) {
-    g_cleanup_interval = config->server.cleanup_interval * MS_PER_SECOND; // Convert to ms
+    g_cleanup_interval = (uint64_t)config->server.cleanup_interval * MS_PER_SECOND; // Convert to ms
     platform_log_debug("Cleanup interval set to %d seconds\n", config->server.cleanup_interval);
   } else {
     platform_log_warning("Invalid cleanup interval %d, using default 5 seconds\n", config->server.cleanup_interval);
